Add Group ID to create response for later use
If we create a group and later want to make a change, we'll need the ID.
Since groupadd doesn't return anything other than an exit code, we use
posix_getgrnam to fetch data of the created group.

On *nix, `getent group <name>` would be the equivalent. `getgrnam` is
called behind the scenes but, at least on WSL, this doesn't exist.

Besides, using the posix function saves a dodgy call to `exec()`.