Fix whopping great whitespace in user/group lists
When the editors have a longish list of users/groups shown causing it to
be longer than the main list, everything gets spaced out weirdly due to
the 'stretched' class. Wrapping it in a row helps to an extent, but
still doesn't do the job properly. We'll just scrap stretched mode.