Resolve #38 - Add info if user/group list is empty
When a group has no users, or a user has no secondary groups, there's no
list and nothing to say a list would be there, just a button. This adds
an info message letting the user know that it's not broken, there's just
nothing to see 'til they add a user to the group or a group to the user.