Fix fallback routing
With the previous setup, the catch-all worked fine for frontend routes
but also included routes with the /api prefix, despite not being defined
in routes/api.php.

To resolve it, this adds an extra catch-all to the API routes, returning
with a 404. Any non-API routes that don't match will still be passed
through to Vue so that vue-router can handle them.

As an aside, Route::fallback() got added in 5.5.5 and does the exact
same as we did originally, with a catch-all on GET routes. It doesn't
affect non-GET routes though, so it'd be no good for the API side.