Move statsbar data to API route
Since requiring auth, we no longer want the stats bar data to be
included in every view, so this moves it to an api route instead.

If the user isn't authed though, attempting to display it results in
some ugliness whereby the orange blob with linux icon would display with
no text around it, so we also add a `loaded` property to ensure it's
only displayed when - and ONLY when - the data is available to fill it.