Bump minimum PHP to 7.3
Travis is complaining that our trailing comma in a function somewhere is
causing a syntax error. Rather than fixing it, let's force PHP 7.3 for
both Travis and our composer.json to make sure people meet the minimum.

Let's face it, 7.0 is dead, .1 is already out of active support and 7.2
will be following in its footsteps before the end of the year anyway.