Add basic site "editor" page
It doesn't actually edit anything yet, but this at least gets the route
etc in tact ready for edit functionality to be added later on.

The edit route takes `id` as a param which is checked to be numeric and
cast to an int so that we can use === to find the site from the store.

In the Editor component we have to give a default value for the site
object, or it simply doesn't work. Why? Fuck knows. But Vue bitches
about it being undefined if we don't so let's just keep it happy. For
some reason it also requires arrays and objects to be defined from
functions, which seems dumb but hey-ho. Instead of having an ugly mess
like `default: function() { return {}; },` we use the `() => ({}),`
shorthand. Does the exact same thing but with confusing syntax.

It would be nice if we could send a 404 if ID was non-numeric or less
than zero, but for now we just default to 0 and break the edit page.