Add dummy fields for project/nginx config
Creates a new section for the primary domain containing various settings
related to the nginx config which vary based on the type of project,
based on the template fields in #58.

Also expands the form in the edit page to use the full remaining width.

Still need an area for custom nginx blocks, but most of #4 is done here.