Update new fields to use main tmpSite object
With `return {...this.site}` in the computed `tmpSite` property,
changing the dropdown had no affect after switching the inputs to use
`tmpSite` due to `this.site` essentially being re-cloned every time
something changed. Didn't seem to be an issue before, but then there
were only one or two inputs previously.

Attempts to clone `this.site` in `created` and `mounted` events failed
as `this.site` wasn't actually loaded yet, so this is a bit of a hack to
workaround it. We use a new data attribute `clonedSite` to keep a copy
of the original `site` object and check if it's empty - which it will be
until the `loadSites` action has finished doing its thing when triggered
from the parent `Sites.vue` layout.

To avoid data going stale (ie not being re-cloned) when we try to edit a
different site, we also add a check to make it clone when IDs differ.