Set document_root on the (cloned) Site object
Using :value isn't going to push its value into `tmpSite`, so instead we
set it from a new `setDocroot` method triggered when either the primary
domain or website type change.

The latter is achieved with a watcher as Semantic UI's dropdown
component doesn't trigger @change events like a standard `<select>`
would (i.e. it doesn't trigger *at all*). This has the added bonus that
it's triggered when `this.sites` is initially populated on page load so
we don't need to call `setDocroot()` from the `tmpSite` computed prop.