Load sites in `beforeRouteEnter`
Moves `tmpSite` into the VueX module and dispatches the `loadSites`
action from `beforeRouterEnter` rather than `mounted` so that the sites
array will always be populated before trying to clone the site object.

Downside is the temporary object etc. is no longer contained within the
Editor component, but on the other hand now that the `loadSites` action
returns a `Promise` it'll be much easier to add 404 handling in future.

`this` isn't available in `beforeRouteEnter` (the component isn't loaded
yet, so it makes sense) meaning we have to manually import the `store`
instance. Bit ugly, but it works. Because we have to use `to.params` to
access the Site ID, it's not being converted to int by our prop handler
in the `routes.js` file either, so yay for extra hacky goodness.

Partially reverts 61eb168. Since moving tmpSite into VueX module, the
`sui-checkbox` component started bitching about getting a `Number` even
after running the `.toString()` function on it in `setEditorSite`.