Fix #57 again, this time with a test
Turns out id is the wrong thing to use, must've been a fluke when it
worked with the original commit that "fixed" the issue.

Adds a new test to make sure it stays working and moves the other update
test to the bottom so that create and update tests are kept together.