Ensure guests cannot access/modify sites
Moves the sites resource route within the `auth:api` middleware group to
ensure no unauthenticated requests can touch them, and updates tests.

Also updates the two FormRequest classes to return false in `authorize`
if a user is not logged in. We shouldn't ever hit this, but it's a good
safety measure just in case some other thing ends up using it later.

Fixes #55.