Add vagrant setup
Adds configuration files for a vagrant environment so we have somewhere
consistent to start working on Nginx features without needing to mess
with the host machine or fire up a droplet.

Could do with improving the .zshrc somewhat, but this works. Also
updates the example .env file to reduce amount of sed needed in
Vagrant's provisioning script.