Fix vagrant shared folder permissions
Sets mount options on the shared folder to make sure files get created
with 664 permissions, ie a+rw,g+rw,o+r, and updates the composer command
to run with the vagrant user using its own home directory so that it can
cache packages into ~/.composer/cache.

Normally we'd just use ACLs (with `setfacl`) to solve permission issues,
but they don't work with shared folders inside a VM so that's a noop.