Save nginx conf for sites when they change
Adds Blade templates with Nginx configs for the Basic, PHP and Redirect
site types. The parsed views will first be saved locally before being
copied using sudo into /etc/nginx/sites-available.

While still very basic, this is a start for server templates in #59 and
enough to finally close #58 now that blade templates for vhosts are in.

Also closes #58 and #62,