Create symlink when site is enabled (#63)
When a site is updated, if it's enabled and a symlink doesn't already
exist with the same name, we now create it before restarting nginx.

Still need to cleanup when disabled and handle real files when a symlink
is expected, but this should be enough to get it working for now.