Add switch to change Sites' is_enabled flag
Moves the form tag to be above the header so that we can jam the
is_enabled toggle switch (technically a checkbox hence needing it in the
form) between it and the header. That way we can float it to the right
and it's kept out the way of the rest of the stuff.

Fixes #85.