Properly maintain symlink in sites-enabled
Does a better job of creating the symlink by ensuring any existing files
are removed prior to running `ln -s`, and skipping the process entirely
if the link already exists and goes to the right place.

Also adds handling for the removal of symlinks when (and only when) a
site is disabled. We avoid removing the file if it's not a symlink or it
is but goes elsewhere just in case this site hasn't yet been enabled and
there's still another manually created site that the config is for. Once
it's been enabled it's safe to assume the old site is being replaced.

Closes #63.