Actually pull the source repos into docroot
Instead of creating the docroot dir itself, we now create its parent so
that Git can clone the site into it. If there's already a .git dir in
the docroot then we'll just pull, but if it's a redirect we do nothing.

Also fixes the test for updating sites, since a non-existent repo causes
user/pass prompts in the shell throwing a fatal error.