Validate git branches
This code is FAR from pretty, and I would've much preferred to have it
in its own Rule class, but apparently those don't get fed the validator
or any of its data, so it's pretty useless given we need other fields.

Still, it works. We pass --exit-code to ls-remote so that it spits out a
status of 2 if there are no branches, which should make it easier to
troubleshoot vs any other given status codes.

Refs are returned from ls-remote as `<hash> refs/heads/<name>` so rather
than messaging about splitting on \t we just go straight for substr,
offsetting by the length of "refs/heads/".