Simplify branch checking
Probably should've thought sooner but turns out there are flags on
ls-remote to filter by heads or tags, so we can just add --heads and
search for the specific branch rather than grabbing them all and
searching the array of branches after the fact.

Cleans things up a little and gives a bit more meaning to the "no
branches" response, but it's still not the prettiest thing.