Open Site Editor when search has exact match
This isn't ideal given it's unclear what will happen when you press
enter. A dedicated create button would be nice so this can just be a
search box, but we don't have a create form for sites.

Since we're looking for an exact match, `find` will only ever return an
object if successful, or undefined when it can't find anything. We could
check against it being undefined, but that could lead to problems if we
end up switching to a more fuzzy (or inclusive) search later on.