Set icon based on search text
This updates the filter input's icon to start off as a magnifying glass.
If an exact match is found it'll change to a cogs icon to match the
'Manage Site' icon in the site list. In all other cases, it'll be a plus
icon indicating that pressing enter will create a new Site.