Fix hot module reloading and remove Browser Sync
I'd just given up on getting Mix to work with HMR (specifically with
CORS since we're using servidor.local which points to Vagrant, and HMR
hosts the assets on localhost). Turns out instead of trying to set the
allowed origins header we need to disable the host check.*

We no longer have a need for browser-sync now either, hence its removal.

Also cleans up the npm commands. Don't need two of everything and we
don't need the full path to webpack[,-dev-server] either.

* See webpack/webpack-dev-server#1604 for details