Add makefile to make testing easier
Because we're using vagrant, we could do with running tests inside the
VM to make sure things like Git commands work as expected and to save
having to mess about maintaing two sets of database options for testing.

This makes it so we can just run `make test` and it'll automatically SSH
into the VM, change to the right dir and finally run tests as www-data.

We could just add a script to package.json and use `npm run test`, but
that has a LOT more garbage output and creates unnecessary logs too.