Replace dummy file list with real results
Adds a new API controller for listing the files/folders of a given
directory, along with the usual bootstrapping stuff to get it loaded up
in Vue with its own VueX store etc.

Permissions could do with not being in Octal, much like owner and group
would rather not be the IDs, but this'll do to get some actual files in.