Add "up one level" button
This commit adds logic to keep track of the current path and uses that
in the frontend in place of the actual path prop, which is now only used
to set the initial path of the file manager.

With the current path being tracked we can just chop everything off
after the last slash and use that to go up a level.

Maybe at some point in the future we could log *all* paths as we
navigate through the tree, allowing for going back and forward rather
than just up (which will be useful when symlinks point elsewhere).