Add ability to view file contents
Removes the old 'Files' page component since it wasn't really doing
anything, instead replacing it with the browser component which now has
its default path set directly from the routes file.

That was originally to make it easier to use child components in the
routes, but I ended up not doing that at all. Either way, with a bit of
hackery in Laravel's routes thanks to Taylor being Taylor, we can now
point our browser to /files/edit//path/to/some/file.txt to view it.

Ideally it would only work when the slashes of the path are actually
encoded, but again, Taylor does what Taylor wants, so `/` and `%2F`
might as well be the same goddamn thing despite being VERY different.

We're just dumping the file contents out in a `<pre>` for now.