Add back button to return to directory listing
Updates the main listing route to accept an optional path parameter in
the url so that we can use that if it's set, falling back to /var/www.

We have to move the index route below the editor otherwise the editor
links will all be seen as part of a path intended for /files.