Fix double / at start of paths after root dir
If you go up a level until you reach the root of the filesystem, it
lists the files/dirs just fine but all paths you navigate to after that
end up getting prefixed with an extra slash.

If we're in the root dir, we can just omit currentPath to fix it.