Fix test for travis
Updates file/dir paths to ones that'll exist on pretty much every linux
box this thing might run on. Hopefully the travis image doesn't purge
the skeleton directory...

Dummy structure in the repo would be better in the long run (or creating
one on the fly in setUp()) but this should be enough to get it passing.