Add value tests for file listing
This makes sure that (some) values are actually what we expect,
something we couldn't really do easily before due to the sort order
being somewhat random and unpredictable.

Not sure what user/group Travis is running phpunit as, so for now we're
leaving the owner/group assertions as simple key checks just in case.