Switch test-skel owner to www-data for consistency
Before now we couldn't test the values of the owner and group keys
because they'd be different in Travis than they were in Vagrant.

Travis by default runs everything as `travis` with the source repo
cloned under travis:travis, whereas Vagrant would be vagrant:www-data.

This tells Vagrant to use www-data for its shared folder and explicitly
sets the owner and group of our test skeleton structure in Travis since
git is only capable of tracking the permissions. I'm sure there was a
reason we initially used the Vagrant user as owner on the synced folder,
but tests are still passing and everything still seems to work.