Add release script and servidor:version command
This moves version definition to bootstrap/app so that it can be
referenced by the Artisan command, which is in turn used by our release
script to check the current version so that it knows what to replace
when we bump to a new version. Should work, but we'll find out soon!

Only real difference from before is that we now have to ensure it's not
defined because bootstrap/app.php is called multiple times in testing.