Add corresponding button to the File Manager
Adds a new button to the top right of the file manager next to the path.
It'll detect if the current folder matches the document root of any
sites and, if it is, open the Site editor to the given site's ID.

Would be nice to have this work in subfolders too, but that's a lot more
complicated than it first sounds so it'll have to wait 'til later on.