Handle errors when opening file without permission
When Symfony's SplFileInfo fails to load a file with file_get_contents
it will pass on the error as-is in a RuntimeException, which we now
catch and parse to modify the error message in certain cases.

If the error is not one already handled (currently limited to Permission
denied errors) then it'll fall back to the exception's message with a
418 status code on the response.

To make sure that tests succeed on the permissionless file, we also need
to nest an extra synced folder within the main one so that we can
override the owner and permissions on the files it contains.