Check for and format errors on the frontend
Adds a placeholder segment so that errors can be displayed in a somewhat
nicer format with a big coloured icon above them.

When we're handling the RuntimeExceptions thrown by Symfony, we want to
make sure the full response data is saved as it's already got everything
we need. We still need to handle other server errors though, so for
those we create a minimal object that gets saved with the error instead.