Handle "Malformed UTF-8 characters" error in files
When attempting to load images and likely other binary file types, the
server throws a 500 with "Malformed UTF-8 characters, possibly
incorrectly encoded" in an InvalidArgumentException.

I'm sure there are other errors this exception covers, but this at least
cleans it up a bit. Logging in the future of the original error detail
(without passing it to the client) would be nice to catch false
positives. This goes some way to solving #103, but not quite.