Make domain name required when updating Sites
Requiring it for the create request would be a terrible idea given we
only have a name box on the frontend, heck it'd be nice to not require
it at all which would enable "draft" sites that have no domain yet.

Until we add a check to the enabled toggle to make sure the domain is
set though, we'll have to make sure it's required. If we don't, you can
almost guarantee Nginx will spew a bunch of errors with blank domains.