Add free RAM to StatsBar
I wanted to show "<free>% of <total>" but SemanticUI's statistic values
are far too big, and if I switch the value and labels the icon looks
absolutely ridiculous, so just the Free RAM will have to do for now.

Perhaps we can add used/total after a bit of a redesign of the statsbar.