Replace current path text with breadcrumbs
This reverts the 'up' button back to normal size since the breadcrumb in
its large size seems to work quite well with normal sized buttons.

To make sure we don't get an accidental trailing slash everywhere, we
check that we're not on the last index while iterating through the parts
of the path. By using a template tag, we don't need a wrapper div.