Link each path segment to its respective folder
This isn't perfect, the spacing is slightly off to the right of the
breadcrumb separators because we're wrapping it in a link rather than
making the crumb itself be the link.

Problem is, if we don't do it this way, there'll be nothing to click on
for the root element as only the label itself can link, not the divider.