Fix route not updating
This restores the template-based markup with click handlers instead of
setting the href. It wasn't loading the files because the component
wasn't actually changing between paths.

Since we need to load the files in both the mounted *and* before the
route upates (otherwise it won't first load until you go up a dir), we
may as well go all out with the click handler too.

To solve the issue of not being able to get to root, the divider is also
made clickable although it isn't particularly obvious in the UI.