Clean up appearance of path bar
This sets dividers to share the same colour and cursor as the rest of
the clickable parts, and adjusts the margin such that a small amount to
the right of the '/' is also clickable.

The left side of the first (root) divider is also pushed out a fair
chunk to the left for a much wider target area to click on.

With this adjustment to margins and padding, it's also spaced out
correctly and doesn't look odd from a distance any more. Bonus.