Don't link the current/final path segment
We could check if the index is one less than the total number of path
segments in the array, but it seems more readable to check whether it
matches the current path and does the same thing.

As it's no longer a link, the breadcrumb section component spits the
last segment out in a div, so we've removed the 'a' limit from the css.