Merge the two path-setting functions
Since we're passing either the absolute path as a string, or a complete
file object containing its path, we can clean things up by consolidating
both into one function that simply sets the path based on input type.

I fiddled with one long ternary to set the path, but it was a bit on the
unreadable side, so instead I opted to settle for a single if/else.

This also fixes an issue introduced when we added the "route update"
handler where the loadFiles event would be dispatched twice when
browsing by clicking on the folder rows.