Move user menu to bottom of the sidebar
The popup type thing we were getting from the dropdown looked like
complete and utter crap, and didn't work if I tried changing its side.

This converts the dropdown to an actual menu, with the idea that we can
animate it manually soon so that it slides up/down when toggled.