Set user menu height with CSS transitions
Reverts most of the previous commit and uses CSS to animate setting the
max-height as the cursor is hovered over or moved away from the menu.

While I'm not always the biggest fan of hover menus, this seems to work
well so far (though can't vouch for mobile...) and saves messing with
hover/active classes too.