Merge System menu into the main sidebar menu
Removes the old System layout and menu components and updates the
users/groups pages to be standalone rather than children of a /system
route, since we don't want to end up with a duplicated layout.

Still isn't perfect, seems to be a bit of spacing below the submenu so
that could do with being re-done at some point, but meh.