Fix padding of main content section on some pages
The file browser and System users/groups all looked fine when the window
was over a certain size, but shrink it down enough and they'd lose their
padding making it look a bit messy.

Everything directly within a grid container is padded by 1rem at the
sides, so this updates the components to use them. We also need to wrap
it all in a grid column, otherwise padding won't apply to all elements.

The Sites page is absolutely fine, but since the grid component has a
proper `container` prop, we'll use that for consistency with the others.