Update tests for user and group lists
Moves the List Users test into a new Users folder ready for adding tests
for create/update/delete later, and adds a few new tests/assertions to
bring user and group list tests in line with each other.

Also updates the tests to depend on the main 'can view' test so we don't
have to fire a request for every test.

When we're testing for a normal user it's a bit difficult since every
distro/install is likely to be using something different (especially if
we're running locally but outside vagrant) so we pull some magic off to
test for a specific user in Travis/Vagrant, then falling back to
whatever user PHP is running as. Not ideal, but until I can figure out
some way of making the lists return data from /test-skel/etc, it'll do.