Fix some groups being left behind after tests
'newtestgroup' and 'updatetestgroup-renamed' were not being deleted
after running tests, which meant on the next run any tests trying to
create those groups would inevitably fail.

This would obviously be much better if we defined the deletion bits in
some parent class or trait, but that's for after user tests are done.