Fix huge amount of whitespace above most pages
All pages using the grid as its root element (that is, everything except
Dashboard) had a whopping amount of whitespace between the Stats Bar and
the search box since the redesign.

While there are likely better ways to fix it, the easiest seems to be
reducing top margin and switching the Dashboard to use the same grid as
other pages. We keep the text class for now since it sets font size,
among other things (such as max width) which won't be as important once
we get some actual content to show on the dashboard.