Normalise spacing between Site search and content
While the sites list has the same gap as most other pages between the
search/create bar and the actual list of Site cards, the editor has an
extra 1rem of padding due to the .row which drives me insane.

The 'Editing <site>' header also has a huge top margin which pushes the
entire edit column down, offsetting it from the top of the list of Sites
on the left and making it look off.

Both these issues are fixed by adding an id to the main Sites content
section since we can't add one to the grid container itself (it would
get replaced by #content) then we can apply rules only to Sites.