Fix redirect to login on 401 Unauthenticated
When an AJAX request returns an HTTP 401 indicating that the token is
incorrect or expired, it's supposed to redirect to /login to force
reauthentication, but doesn't actually remove the token.

Problem is the login route requires guest -- that is, it checks to see
if there's already a token set in the store and, if so, redirects to the
Dashboard on the assumption that the token is valid.

This updates the error handler so that it clears the token before
pushing the new route. In addition, we make sure to clear the token from
local storage and moves the setting of it into the actual mutation. This
way we'll know it's always kept in sync.