Force re-login when token changes mid-session
While unlikely to happen often, it's plausible that occasionally the
auth token in local storage could be changed or removed such that it no
longer matches the one stored in the current VueX state.

This is most likely to happen if a user clears caches / local storage,
but when it does we should force a new login just in case. We don't want
the tokens to differ in case both happen to be valid and actions end up
being performed under the wrong credentials.