Set up PHP Mess Detector
Adds phpmd to the travis config with a custom ruleset that uses most of
the standard rules with 5 exceptions. StaticAccess is pretty much
required on the ignore list given the nature of Laravel's facades.

Also fixes a whole host of warnings it flagged, most of them are pretty
basic stuff but one or two needed minor refactors of code to comply.