Remove composer self-update task
Turns out Travis has been doing it all along before we even get to the
install stage, so running self-update will always be pointless. The
validate command is kind of implied by installing from composer.json
too, so while we're at it we might as well ditch that as well.