Add site details page with pull and browse buttons
Creates a new 'overview' page with details about a project's git
repository and document root (or redirect target/type) along with
buttons to either browse the code or pull the latest changes from Git.

The git stuff is fairly close to being a complete duplicate of what's
already in Listeners\WriteSiteConfig so it probably needs a refactor.

Closes #91.