Fix Travis build errors
The first fix is of a complaint from phpmd about having an else clause
in the Site controller's `pull` method. Slightly annoying having to exec
and return essentially the same thing twice, but there ya go.

The second fix is that of a permissions issue in Travis itself - which
is to say it works fine locally. For some reason it can't create the
`www` dir inside resources, but it's fine in storage or anywhere else.