Fix a bug that caused ':' to appear as a group
Since the switch to Vagrant with an Ubuntu VM, the list of secondary
groups for a user has had an extra ":" group at the top. This in turn
was also causing other issues, such as #154.

PR #165 worked around that by filtering any ':' matches from the groups,
but if we pipe the `groups` command through sed, we can kill two birds
with one stone and eradicate that rogue colon completely.

This is all caused by output that differs between Linux distro, so I
have no idea if any other distros might be affected. Arch output is
something along the lines of `group1 group2 group3`, whereas on Ubuntu
you get `username : group1 group2 group3` - where 'username' will also
likely be the first group in the list.