Clean up codemirror toolbar in the file viewer
This replaces the toolbar segment with an attached menu, adds 'item'
classes to the dropdown buttons so that they're styled correctly and
removes the padding from the actual code editor so it's seamless.

Button dropdowns are a bit off with the item class for some reason. To
fix the icon hugging the left border we need to reset the width from
auto, but that also restores the button's background so we force that to
be transparent instead. That just leaves a left padding tweak to fix the
massive spacing between the button's icon and currently selected value.

This also fixes an issue where there were two loading spinners shown by
removing the loading state from the actions toolbar.