Add branch dropdown to site editor
This replaces the branch input with a hard-coded dropdown that lists the
standard master and develop branches as a start to implementing #92.

Doesn't actually use or set the site's actual source_branch field yet,
we'll need to add a specific getter/setter type mapping that gets from
the getter but actually sets the site.source_branch field instead.