Load branches from Git on site edit
Adds a sites/:id/branches endpoint to fetch an array of Git branches for
the current repository set for the given site, using git ls-remote.

With this, the branches dropdown is now functional in that the branches
are accurate as long as the clone url is set before being loaded. For
new sites, the branch URL will still need to be saved and refreshed.