Update branch list when source repo is changed
Adds support to the branches endpoint for specifying a custom `repo`
query string parameter which will be used in place of the Site's current
source_repo value when it's passed.

Also adds a dedicated loadBranches action that's called both when the
editor is loaded *and* when the source_repo input is changed, with the
difference that repo is only passed when called via the change event.