Add dark mode
Adds basic support for the Superhero theme ported from Bootswatch, with
custom patches applied to make the stats and sidebar blend in better. To
support both light and dark themes, semantic-ui and the layout styles
have also been extracted to a new core.scss file so it can be loaded
before the rest of the theme and other app styles/overrides.

Theme switching support is added by way of a plugin that provides toggle
component and mixin with the various data and methods which all come
together to make setting and getting the theme possible, with the
default theme set to the previous "light" mode.

This commit comes with a caveat:
Since moving the app and theme related styles to their own files and -
perhaps most importantly - calling them without `mix()`, hot reload
hasn't been working all that well. For now we need to npm run dev, which
is quite frankly a bit of a ballache while working on vue or sass stuff.