Update user and group pages for dark mode
This adds the inverted class to the search and edit forms to take
advantage of tweaks added previously, also inverting the danger segments
and the button contained within them so they don't look so out of place.

The subheader rule is tweaked so that it doesn't require an h2, which
means it'll now apply to the sub header in "no results" placeholders.

Finally, the sui-grid-column components get a userslist or groupslist
class so that we can group the extra custom styles, which fix:
* Dark text in the search box while it's focused
* "Show system" toggle label (!important to prevent :checked override)
* Background of user/group member list in editors
* Segment "container" borders and backgrounds (which were white!)
* Active icon being changed to black in main lists
* Glitches when hovering main user/group list items while in editor