Fix db and file manager's appearance in dark mode
The file list and database table were just a case of setting the
inverted flag, with the database search box taking advantage of the
styles added earlier for transparent inverted inputs.

Purple's a bit difficult to see on dark grey backgrounds, so this
changes the database icon to orange in dark mode.

File editor also needed a tweak to loosen the label rule in the css,
otherwise the 'Line Wrapping' label would be black. Otherwise, like the
other views this is mostly a case of adding the inverted prop to things.