Fix tests
If we don't have permission, we'll catch that exception when it tries to
load the content so we can suppress the warnings that PHP throws if it
fails to load the mime type of a given file. We also make sure the mime
type was actually set before comparing it to ensure errors are caught.