Add ability to save files
This updates the file controller to handle PUT requests, saving the
given file as long as it exists and isn't identical. For this to work
we've changed the axios request to use the query string for the filename
otherwise it'd make it difficult to add rename functionality later.

We also switch away from resource routes for the file controller so that
we can more easily define our custom routes later.