Fix missing file path on reload
If you were to browse to a file then open it from the file browser,
you'd be able to save it as the path was set by the initial loading of
the parent directory when you were in the file browser. If you then
refresh the file editor though, that path is lost.

This updates the `FileManager` to set the file path in the array it
returns so that we can then use it to setPath in the VueX action.