Fix indents and sort out codemirror config
While tabSize was set to 4, that only affects the spaces displayed for
an actual tab. To get 4 spaces when we *hit* tab, we need indentUnit.

Also enables autofocus and sets a 75px scroll margin so that 3 or 4
lines will always be visible beyond the cursor while moving around the
document - that is, unless you're at the top or bottom.

Options are now sorted according to docs, making it easier to compare.