Move up/back handling into PathBar component
Both the upOneLevel and backToDir methods were essentially doing the
exact same thing, just with a different path. That path is passed along
to the PathBar component anyway though, so we can merge them into one.

I initially had the router.push calls in both goTo and goUp, but since
they're calling the same route we can just call goTo from goUp with the
right path. The original handling (to default to /) is the same in goUp.