Move listing code to SystemUser class
While this still isn't perfect, the index method can largely be moved
as-is, along with handling for secondary groups which it depends on.

The latter is implemented as a method on the user class itself (rather
than static) so this also updates the update method to first create a
user instance from the arrays we have available.