Refactor System\UsersController@update method
This moves most of the update functionality into a new method on the
SystemUser class, replacing `posix_getpwuid` with a call to `::find()`
so that we have something to call the `update()` method on.

Since we're now operating on the `SystemUser` object, we can compare
against `$this->user` instead of an `$original` array. `$newUid` can
also be renamed to `$uid` as the original is in `$this->user['uid']`.

Loading secondary groups is now done via a `loadSecondaryGroups()`
method, which avoids creating an unnecessary new user instance. That
same method is later called again after performing a `refresh()`, which
will run `posix_getpwuid()` but actually save it to `$this->user`.

When it comes to returning a response, we don't just want the user
object because that'd mean making `user` public, then you'd have ugly
stuff like `$user->user['name']`. To enable getting the actual user
data easily, we also add a `toArray()` method. Most others will return
either a new instance or a reference to `$this`.

If a user is not found or modified during update, we now throw a custom
exception which is caught in the controller to fail with a more human
error message. In addition, the `fail()` method will now directly throw
the exception rather than returning and throwing it from update(). A
third exception is also thrown when creating or updating a user fails
(ie the exit status is anything other than 0), which is left uncaught.