Update php-cs-fixer to 2.16.2 for PHP 7.4 support
This is a bump of two major versions, so there are a lot of new rules.
There are only a couple that change things currently, while other rules
added since 2.14.2 can be considered for addition in a later commit.

The `array_syntax` rule is now part of the `@symfony` group, so it has
been removed from our configuration. Fixes applied in this commit are
for `ordered_imports`, `single_trait_insert_per_statement` and also
`no_superfluous_phpdoc_tags`, all of which existed previously but were
added to the `@symfony` group in either 2.15 or 2.16.