Abstract system user validation
This moves the validation rules into a new SaveUser request class, with
the custom callback for 'name' being further separated into a new Rule.

This has the downside that specific invalid characters (eg comma) are no
longer shown in the error messages, but that is likely to be obvious.
We're already using a regex to catch other things anyway.