Update PHPStan to 0.12.0
In order to use 0.12, larastan also required an update but the PR to
support 0.12 of phpstan is currently on hold due to an error emitted
from the Console command. As I don't use that, this uses my own fork
of that PR, removing the Console command entirely to fix the issues.

When nunomaduro/larastan#378 is merged, the `repositories` block will
need to be removed from composer.json so that we can go back to using
the main Larastan releases. Version requirement should also be reset.